Fix how we were doing the filtering of web pages. Divide the url in url_path and domain. And simplify the blacklist.
The previous filter len(url.split(".")[-1].split("/")[-1]) did not capture all the homepages. For example the webpage 'https://www.test.com' would not be catched. Also 'http://www.test.com/otro.html' will not be catched New filter is to check if the last char is '/'. Which should catch homepages and folders.